+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.0] - Under development
+
+### Added
+
+
 ## [0.2.0] - 2023-09-15
 
 ### Added
@@ -21,5 +26,7 @@
 - Merged results can be displayed in tabular output, CSV, or in interactive terminal browser.
 - Interactive browser will use the screen width by default, or accept a command-line argument.
 
+
+[0.3.0]: https://github.com/ptmcg/log_merger/compare/v0.2.0...main
 [0.2.0]: https://github.com/ptmcg/log_merger/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.1.0